[WEB][UPDATE]: Refactored Vite configuration to conditionally enable sourcemaps and set build target to ES2020. Enhanced chunking strategy for vendor libraries. Updated query client configuration to utilize constants for stale time management, improving cache handling. Introduced lazy loading for various components to optimize performance. Adjusted image loading attributes for better user experience. Added constants for query cache management to standardize stale time settings across the application.
