+++ apps/web-seller/src/common/constants/paths.constant.ts
@@ -1,32 +1,96 @@
 export { ROUTES } from "@/apps/web-seller/common/constants/route-paths.constant";
 
 import { ROUTES } from "@/apps/web-seller/common/constants/route-paths.constant";
-import { RootPage } from "@/apps/web-seller/pages/Root";
-import { StoreCreatePage } from "@/apps/web-seller/pages/store/Create";
-import { StoreDetailHomePage } from "@/apps/web-seller/pages/store/detail/Home";
-import { StoreDetailProductListPage } from "@/apps/web-seller/pages/store/detail/products/List";
-import { StoreDetailProductCreatePage } from "@/apps/web-seller/pages/store/detail/products/Create";
-import { StoreDetailProductDetailPage } from "@/apps/web-seller/pages/store/detail/products/Detail";
-// 채팅 노출 시 복구
-// import { StoreDetailChatListPage } from "@/apps/web-seller/pages/store/detail/chat/List";
-// import { StoreDetailChatRoomPage } from "@/apps/web-seller/pages/store/detail/chat/Room";
-import { StoreDetailEditPage } from "@/apps/web-seller/pages/store/detail/Edit";
-import { StoreDetailCalendarPage } from "@/apps/web-seller/pages/store/detail/Calendar";
-import { StoreDetailFeedListPage } from "@/apps/web-seller/pages/store/detail/feed/List";
-import { StoreDetailFeedCreatePage } from "@/apps/web-seller/pages/store/detail/feed/Create";
-import { StoreDetailFeedDetailPage } from "@/apps/web-seller/pages/store/detail/feed/Detail";
-import { StoreDetailOrderListPage } from "@/apps/web-seller/pages/store/detail/orders/List";
-import { StoreDetailOrderDetailPage } from "@/apps/web-seller/pages/store/detail/orders/Detail";
-import { StoreDetailNotificationsListPage } from "@/apps/web-seller/pages/store/detail/notifications/List";
-import { StoreDetailNotificationsSettingsPage } from "@/apps/web-seller/pages/store/detail/notifications/Settings";
-import { StoreDetailStatisticsPage } from "@/apps/web-seller/pages/store/detail/statistics/Index";
+import { lazyPage } from "@/apps/web-seller/common/utils/lazy-page.util";
 import { LoginPage } from "@/apps/web-seller/pages/auth/Login";
-import { FindAccountPage } from "@/apps/web-seller/pages/auth/FindAccount";
-import { GoogleAuthCallbackPage } from "@/apps/web-seller/pages/auth/GoogleAuthCallback";
-import { KakaoAuthCallbackPage } from "@/apps/web-seller/pages/auth/KakaoAuthCallback";
-import { MypageIndexPage } from "@/apps/web-seller/pages/mypage/Index";
-import { TermsOfServicePage } from "@/apps/web-seller/pages/terms/TermsOfService";
-import { PrivacyPolicyPage } from "@/apps/web-seller/pages/terms/PrivacyPolicy";
+
+// 로그인은 첫 진입점 — static import 유지
+const FindAccountPage = lazyPage(
+  () => import("@/apps/web-seller/pages/auth/FindAccount"),
+  "FindAccountPage",
+);
+const GoogleAuthCallbackPage = lazyPage(
+  () => import("@/apps/web-seller/pages/auth/GoogleAuthCallback"),
+  "GoogleAuthCallbackPage",
+);
+const KakaoAuthCallbackPage = lazyPage(
+  () => import("@/apps/web-seller/pages/auth/KakaoAuthCallback"),
+  "KakaoAuthCallbackPage",
+);
+const TermsOfServicePage = lazyPage(
+  () => import("@/apps/web-seller/pages/terms/TermsOfService"),
+  "TermsOfServicePage",
+);
+const PrivacyPolicyPage = lazyPage(
+  () => import("@/apps/web-seller/pages/terms/PrivacyPolicy"),
+  "PrivacyPolicyPage",
+);
+
+const RootPage = lazyPage(() => import("@/apps/web-seller/pages/Root"), "RootPage");
+const MypageIndexPage = lazyPage(
+  () => import("@/apps/web-seller/pages/mypage/Index"),
+  "MypageIndexPage",
+);
+const StoreCreatePage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/Create"),
+  "StoreCreatePage",
+);
+const StoreDetailHomePage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/Home"),
+  "StoreDetailHomePage",
+);
+const StoreDetailProductListPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/products/List"),
+  "StoreDetailProductListPage",
+);
+const StoreDetailProductCreatePage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/products/Create"),
+  "StoreDetailProductCreatePage",
+);
+const StoreDetailProductDetailPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/products/Detail"),
+  "StoreDetailProductDetailPage",
+);
+const StoreDetailEditPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/Edit"),
+  "StoreDetailEditPage",
+);
+const StoreDetailCalendarPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/Calendar"),
+  "StoreDetailCalendarPage",
+);
+const StoreDetailFeedListPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/feed/List"),
+  "StoreDetailFeedListPage",
+);
+const StoreDetailFeedCreatePage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/feed/Create"),
+  "StoreDetailFeedCreatePage",
+);
+const StoreDetailFeedDetailPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/feed/Detail"),
+  "StoreDetailFeedDetailPage",
+);
+const StoreDetailOrderListPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/orders/List"),
+  "StoreDetailOrderListPage",
+);
+const StoreDetailOrderDetailPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/orders/Detail"),
+  "StoreDetailOrderDetailPage",
+);
+const StoreDetailNotificationsListPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/notifications/List"),
+  "StoreDetailNotificationsListPage",
+);
+const StoreDetailNotificationsSettingsPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/notifications/Settings"),
+  "StoreDetailNotificationsSettingsPage",
+);
+const StoreDetailStatisticsPage = lazyPage(
+  () => import("@/apps/web-seller/pages/store/detail/statistics/Index"),
+  "StoreDetailStatisticsPage",
+);
 
 // 인증 관련 경로 (AdminLayout 밖)
 export const AUTH_ROUTE_CONFIG = [
@@ -52,11 +116,6 @@ export const ADMIN_ROUTE_CONFIG = [
     path: ROUTES.STORE_DETAIL_PRODUCTS_DETAIL(":storeId", ":productId"),
     element: StoreDetailProductDetailPage,
   },
-  // { path: ROUTES.STORE_DETAIL_CHAT_LIST(":storeId"), element: StoreDetailChatListPage },
-  // {
-  //   path: ROUTES.STORE_DETAIL_CHAT_ROOM(":storeId", ":roomId"),
-  //   element: StoreDetailChatRoomPage,
-  // },
   { path: ROUTES.STORE_DETAIL_FEED_LIST(":storeId"), element: StoreDetailFeedListPage },
   { path: ROUTES.STORE_DETAIL_FEED_CREATE(":storeId"), element: StoreDetailFeedCreatePage },
   {
