+++ apps/web-user/src/common/styles/globals.css
@@ -1,3 +1,8 @@
+@import "swiper/css";
+@import "swiper/css/pagination";
+@import "swiper/css/navigation";
+@import "swiper/css/thumbs";
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
