+++ apps/web-user/src/features/home-banner/hooks/queries/useHomeBanners.ts
@@ -1,4 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
+import { QUERY_STALE_TIME } from "@/apps/web-user/common/constants/query-cache.constants";
 import { useQueryErrorAlert } from "@/apps/web-user/common/hooks/useQueryErrorAlert";
 import { homeBannerApi } from "@/apps/web-user/features/home-banner/apis/home-banner.api";
 import { homeBannerQueryKeys } from "@/apps/web-user/features/home-banner/constants/homeBannerQueryKeys.constant";
@@ -12,6 +13,7 @@ export function useHomeBanners() {
       const response = await homeBannerApi.getHomeBanners();
       return response.data;
     },
+    staleTime: QUERY_STALE_TIME.STATIC,
   });
 
   useQueryErrorAlert(query);
