+++ apps/web-user/src/features/product/components/modals/ReviewDetailModal.dynamic.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+export const ReviewDetailModal = dynamic(
+  () =>
+    import("./ReviewDetailModal").then((module) => ({
+      default: module.ReviewDetailModal,
+    })),
+  { ssr: false },
+);
