+++ apps/web-user/src/features/product/components/sections/reservation-bottom-sheet/ReservationBottomSheet.dynamic.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+export const ReservationBottomSheet = dynamic(
+  () =>
+    import("./ReservationBottomSheet").then((module) => ({
+      default: module.ReservationBottomSheet,
+    })),
+  { ssr: false },
+);
