+++ apps/web-user/src/features/product/components/sections/reservation-bottom-sheet/index.ts
@@ -1,2 +1,2 @@
-export { ReservationBottomSheet } from "./ReservationBottomSheet";
+export { ReservationBottomSheet } from "./ReservationBottomSheet.dynamic";
 export type { OrderItem, ReservationBottomSheetProps } from "./types";
