+++ apps/web-user/src/features/product/hooks/queries/useProductDetail.ts
@@ -1,4 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
+import { QUERY_STALE_TIME } from "@/apps/web-user/common/constants/query-cache.constants";
 import { useQueryErrorAlert } from "@/apps/web-user/common/hooks/useQueryErrorAlert";
 import { productApi } from "@/apps/web-user/features/product/apis/product.api";
 import { productQueryKeys } from "@/apps/web-user/features/product/constants/productQueryKeys.constant";
@@ -9,6 +10,7 @@ export function useProductDetail(productId: string) {
     queryKey: productQueryKeys.detail(productId),
     queryFn: () => productApi.getProductDetail(productId),
     enabled: !!productId,
+    staleTime: QUERY_STALE_TIME.DETAIL,
   });
 
   useQueryErrorAlert(query);
