+++ apps/web-user/src/features/store/hooks/queries/useStoreDetail.ts
@@ -1,4 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
+import { QUERY_STALE_TIME } from "@/apps/web-user/common/constants/query-cache.constants";
 import { useQueryErrorAlert } from "@/apps/web-user/common/hooks/useQueryErrorAlert";
 import { storeApi } from "@/apps/web-user/features/store/apis/store.api";
 import { storeQueryKeys } from "@/apps/web-user/features/store/constants/storeQueryKeys.constant";
@@ -9,6 +10,7 @@ export function useStoreDetail(storeId: string) {
     queryKey: storeQueryKeys.detail(storeId),
     queryFn: () => storeApi.getDetail(storeId),
     enabled: !!storeId,
+    staleTime: QUERY_STALE_TIME.DETAIL,
   });
 
   useQueryErrorAlert(query);
