+++ apps/web-user/src/features/store/hooks/queries/useStoreRegions.ts
@@ -1,4 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
+import { QUERY_STALE_TIME } from "@/apps/web-user/common/constants/query-cache.constants";
 import { useQueryErrorAlert } from "@/apps/web-user/common/hooks/useQueryErrorAlert";
 import { storeApi } from "@/apps/web-user/features/store/apis/store.api";
 
@@ -6,6 +7,7 @@ export function useStoreRegions() {
   const query = useQuery({
     queryKey: ["store", "regions"],
     queryFn: storeApi.getRegions,
+    staleTime: QUERY_STALE_TIME.STATIC,
   });
 
   useQueryErrorAlert(query);
